refactor: step3 package에서 step4 package로 변경
