+++ src/main/java/step4/domain/Simulator.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step4.domain;
 
 public class Simulator {
     private final Cars cars;
